+++ src/javascript/app_2/Utils/URL/__tests__/url.spec.js
@@ -1,1 +1,1 @@
-import * as URL from '../url';
+import * as URL from '../url_helper';
