+++ src/javascript/app_2/Utils/URL/base_name.js
@@ -1,0 +1,19 @@
+import { getAll as getAllLanguages } from '../../../_common/language';
+
+/*
+ * Retrieves basename from current url
+ *
+ * @return {string} returns the basename of current url
+ */
+const getBaseName = () => {
+    const regex_string = `(.*(${Object.keys(getAllLanguages()).join('|')})/app(/index\\.html)?).*`;
+    const basename = new RegExp(regex_string, 'ig').exec(window.location.pathname);
+
+    if (basename && basename.length) {
+        return basename[1];
+    }
+
+    return '/en/app/';
+};
+
+export default getBaseName;
