+++ src/javascript/app_2/Utils/URL/index.js
@@ -1,1 +1,1 @@
-export { default } from './url';
+export { default } from './url_helper';
